Refactor ChickenHeadTimer to be in itemabilities
